bit of a clean up
