+++ packages/client/hmi-client/src/components/workflow/ops/dataset-transformer/tera-dataset-transformer.vue
@@ -4,9 +4,22 @@
 		@on-close-clicked="emit('close')"
 		@update-state="(state: any) => emit('update-state', state)"
 	>
+		<template #header-actions>
+			<div class="flex align-items-center ml-auto">
+				<Dropdown
+					v-model="selectedDataset"
+					placeholder="Select a dataframe"
+					:options="Object.keys(kernelState || [])"
+					class="5"
+					:disabled="!kernelState"
+				/>
+				<Button label="Save for reuse" severity="secondary" outlined :disabled="!kernelState" @click="onSave" />
+			</div>
+		</template>
 		<div class="background">
 			<Suspense>
 				<tera-dataset-jupyter-panel
+					ref="jupyterPanel"
 					:assets="assets"
 					:show-kernels="showKernels"
 					:show-chat-thoughts="showChatThoughts"
@@ -14,6 +27,9 @@
 					:notebook-session="notebookSession"
 					:programming-language="node.state.programmingLanguage"
 					@update-language="(lang) => onUpdateLanguage(lang)"
+					@update-kernel-state="updateKernelState"
+					:kernelState="kernelState"
+					:selected-dataset="selectedDataset"
 				/>
 			</Suspense>
 		</div>
@@ -32,6 +48,8 @@ import { cloneDeep } from 'lodash';
 import { v4 as uuidv4 } from 'uuid';
 import TeraDrilldown from '@/components/drilldown/tera-drilldown.vue';
 
+import Button from 'primevue/button';
+import Dropdown from 'primevue/dropdown';
 import { DatasetTransformerState } from './dataset-transformer-operation';
 
 const props = defineProps<{
@@ -49,6 +67,17 @@ const assets = computed(() =>
 			id: inputNode.value![0]
 		}))
 );
+
+const kernelState = ref(null);
+const jupyterPanel = ref();
+const selectedDataset = ref<string | null>(null);
+const updateKernelState = (newKernelState: any) => {
+	kernelState.value = newKernelState;
+	// Default the dropdown to the first dataframe
+	if (!selectedDataset.value) {
+		selectedDataset.value = Object.keys(newKernelState)[0];
+	}
+};
 
 const notebookSession = ref(<NotebookSession | undefined>undefined);
 
@@ -87,6 +116,10 @@ const addOutputPort = (data: any) => {
 		type: 'datasetId',
 		value: data.id
 	});
+};
+
+const onSave = () => {
+	jupyterPanel.value?.openDialog();
 };
 </script>
